Add dom logic for home menu and contact page
In this section, I added the rendering of the home view page to include a brief introduction on what the restraunt was all about. I then went ahead and added the menu section of the page and finished by adding the contact section. All are of which are being rendered by the javasript with respect to the manipulation of the dom