FInish project works
In this commit I have added the dist folder to the eslintignore file since the files in the are automatically generated and it will be useless rectifying the linter errors every now and then. What's more, I have also fixed the other linter errors that was produced in the index.js file and moved the dist folder from the gitignore file for purposes of live review rendering.